DOC: Edit iterator docs
The method links did not render in current nightly, fix the method
links. At the same time, simplify and standardize the docs a bit. Add
docs for the Ones struct too.

The existing links and pointers to where the iterator is created, that's
great, just tweaking the wording a bit. Any Rust iterator is typically
lazy, so we don't need to say that.